EnemyHealth change
- Renamed EnemyBehavior script to EnemyHealth; fixed all references
