+++ Fantasia-Frenzy/Assets/_Mads/SlimeDemon.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 70aa3fb0b8b20f84392cf0884446e075
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
